MAINT Use float64 for molecule positions
Using the default float64 dtype throughout more of the code since it is
no longer necessary to only work with the float32 data type.

+++ src/sdanalysis/util.py
@@ -121,5 +121,5 @@ def orientation2positions(
     mol: Molecule, position: np.ndarray, orientation: np.ndarray
 ) -> np.ndarray:
     return np.tile(position, (mol.num_particles, 1)) + np.concatenate(
-        [rotate_vectors(orientation, pos) for pos in mol.positions.astype(np.float32)]
+        [rotate_vectors(orientation, pos) for pos in mol.positions]
     )
